refactor(api-server): split CRUD handlers
